Fix Result's SelectMany() method.
The intent for SelectMany was to allow chaining multiple results with the same error type but different value types in a LINQ clause with multiple `from` statements. This would short-circuit on the first Err, returning a new result with that error, or an Ok from the `select` statement if all results in the chain were Ok.

The previous version of SelectMany() didn't do that, or even support the indended syntax, so it has been reworked.